# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Risk Control System API",
        "description": "API documentation for Risk Control System",
        "contact": {
            "email": "admin@riskcontrol.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/v1/accounts": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "List all accounts",
                "description": "Returns list of accounts",
                "operationId": "getAccountsList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Create new account",
                "description": "Returns account data",
                "operationId": "storeAccount",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreAccountRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/api/v1/accounts/{id}": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Get account information",
                "description": "Returns account data with risk status",
                "operationId": "getAccountById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Account ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/accounts/{id}/restore": {
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Restore disabled account",
                "description": "Re-enables a disabled account",
                "operationId": "restoreAccount",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Account ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/incidents": {
            "get": {
                "tags": [
                    "Incidents"
                ],
                "summary": "Get list of incidents",
                "description": "Returns list of incidents with optional filtering",
                "operationId": "getIncidentsList",
                "parameters": [
                    {
                        "name": "account_id",
                        "in": "query",
                        "description": "Filter by Account ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "risk_rule_id",
                        "in": "query",
                        "description": "Filter by Risk Rule ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/incidents/{id}": {
            "get": {
                "tags": [
                    "Incidents"
                ],
                "summary": "Get incident information",
                "description": "Returns incident data",
                "operationId": "getIncidentById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Incident ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/incidents/account/{accountId}/stats": {
            "get": {
                "tags": [
                    "Incidents"
                ],
                "summary": "Get incident statistics for an account",
                "description": "Returns incident counts and breakdown by rule",
                "operationId": "getAccountIncidentStats",
                "parameters": [
                    {
                        "name": "accountId",
                        "in": "path",
                        "description": "Account ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
        "/api/v1/incidents/{id}/read": {
            "put": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Mark incident as read",
                "description": "Updates read_at timestamp",
                "operationId": "markNotificationAsRead",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Incident ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreAccountRequest": {
                "title": "Store Account Request",
                "description": "Store Account request body data",
                "required": [
                    "login"
                ],
                "properties": {
                    "login": {
                        "type": "integer",
                        "example": 123456
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "enable",
                            "disable"
                        ],
                        "example": "enable"
                    },
                    "trading_status": {
                        "type": "string",
                        "enum": [
                            "enable",
                            "disable"
                        ],
                        "example": "enable"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}